+++ src/transformers/models/deepseek_v3/configuration_deepseek_v3.py
@@ -179,7 +179,7 @@ class DeepseekV3Config(PreTrainedConfig):
         hidden_act: Optional[str] = "silu",
         max_position_embeddings: Optional[int] = 4096,
         initializer_range: Optional[float] = 0.02,
-        rms_norm_eps: Optional[int] = 1e-6,
+        rms_norm_eps: Optional[float] = 1e-6,
         use_cache: Optional[bool] = True,
         pad_token_id: Optional[int] = None,
         bos_token_id: Optional[int] = 0,
